+++ CHANGELOG.md
@@ -6,12 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Breaking
+- **Deprecated** the `FormLeaveGuardMixin` component in favour of `FormGuardMixin` (fully typed)
+  - _The `FormLeaveGuardMixin` (not truly a mixin...) did not work with TypeScript_
+  - _There was no need for customizing the leave guard to the extend provided_
 - Changed the default behaviour of the `setValues` form function (now will not set initial values by default)
   - _This change was made to align with developer expectations (behaviour moved to `setInitial`)_
+- Removed the form key/name from the `createForm` function API
+  - _This was an unnecessary step that caused more internal work for no gain (simply assign to data)_
 
 ### Added
 - Overhauled package to use [TypeScript](https://typescriptlang.org)!
-- `setInitial` form function to set a form's initial (and current) values (similar to old behaviour of `setValues`)
+- Fully typed `FormGuardMixin` to replace `FormLeaveGuardMixin` (can be customized with `formGuards` data key)
+- New `setInitial` form function to set a form's initial (and current) values (similar to old behaviour of `setValues`)
 
 ## [0.2.3] - 2020-09-30
 ### Added
