+++ src/index.ts
@@ -1,5 +1,8 @@
 import "./classComponentHooks";
 
-export { createForm, FormCreateMixin } from "./formCreateMixin";
+export { createForm } from "./createForm";
 export { FormLeaveGuardMixin } from "./formLeaveGuardMixin";
 export { default as FormGuardMixin } from "./FormGuardMixin";
+
+// NOTE: 'createForm' should be used instead when using TypeScript!
+export { FormCreateMixin } from "./FormCreateMixin";
